+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.4.0
+- Stop setting `box-sizing: content-box` on the input element
+
+## 0.3.1
+- Assign an `id` to the &ldquo;ghost&rdquo; element, and use `document.getElementById` to detect if the element still exists in the DOM
+
 ## 0.3.0
 - Account for case where the &ldquo;ghost&rdquo; element may have been removed from the DOM